# api/bff-openapi.yaml
...
openapi: "3.0.3"
info:
  title: Skateboard UI-backend API
  version: "1.0.0"
...
paths:
...
  /api/config/branding:
    get:
      tags: [branding]
      operationId: getBrandingConfig
      summary: Load the complete branding configuration for the admin screen
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      responses:
        '200':
          description: Branding configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingConfigResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/config/branding/login-text:
    put:
      tags: [branding]
      operationId: updateLoginText
      summary: Update the configurable login screen title and message (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLoginTextRequest'
      responses:
        '200':
          description: Login text updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingConfigResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/config/branding/login-background:
    post:
      tags: [branding]
      operationId: uploadLoginBackground
      summary: Upload/replace the login background image (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Login background updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingConfigResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [branding]
      operationId: removeLoginBackground
      summary: Remove the login background image (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      responses:
        '200':
          description: Login background removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingConfigResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/config/branding/app-logo:
    post:
      tags: [branding]
      operationId: uploadAppLogo
      summary: Upload/replace the app logo image (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: App logo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingConfigResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [branding]
      operationId: removeAppLogo
      summary: Remove the app logo image (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      responses:
        '200':
          description: App logo removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingConfigResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/config/branding/assets:
    get:
      tags: [branding]
      operationId: listBrandingAssets
      summary: List branding assets (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      responses:
        '200':
          description: Branding assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BrandingAssetResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [branding]
      operationId: uploadBrandingAsset
      summary: Upload a new named branding asset (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, file]
              properties:
                name:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Branding asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingAssetResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: An asset with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/config/branding/assets/{assetId}:
    put:
      tags: [branding]
      operationId: replaceBrandingAsset
      summary: Replace an existing branding asset's image (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Branding asset replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingAssetResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Branding asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [branding]
      operationId: removeBrandingAsset
      summary: Remove a branding asset (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_SETTINGS_BRANDING
      parameters:
        - name: assetId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Branding asset removed
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_SETTINGS_BRANDING required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Branding asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/config/home/video-categories:
    get:
      tags: [home]
      operationId: getHomeVideoCategoryConfig
      summary: Get the Home dashboard video-category configuration (admin only — for the config screen)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_HOME_CATEGORY_CONFIG
      responses:
        '200':
          description: Home video-category configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeVideoCategoryConfigResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_HOME_CATEGORY_CONFIG required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [home]
      operationId: updateHomeVideoCategoryConfig
      summary: Update which video categories are eligible for the Home dashboard (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_HOME_CATEGORY_CONFIG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HomeVideoCategoryConfigRequest'
      responses:
        '200':
          description: Home video-category configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeVideoCategoryConfigResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_HOME_CATEGORY_CONFIG required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/config/home/featured-player:
    get:
      tags: [home-featured-player]
      operationId: getHomeFeaturedPlayerConfig
      summary: Get the default Home Featured Player configuration (admin only — for the config screen)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_HOME_FEATURED_PLAYER_CONFIG
      responses:
        '200':
          description: Home Featured Player configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeFeaturedPlayerConfigResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_HOME_FEATURED_PLAYER_CONFIG required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [home-featured-player]
      operationId: updateHomeFeaturedPlayerConfig
      summary: Update the default Home Featured Player configuration (admin only)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_HOME_FEATURED_PLAYER_CONFIG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHomeFeaturedPlayerConfigRequest'
      responses:
        '200':
          description: Home Featured Player configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeFeaturedPlayerConfigResponse'
        '400':
          description: Invalid input (e.g. enabled=true without contentSource/contentId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_HOME_FEATURED_PLAYER_CONFIG required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BrandingConfigResponse:
      type: object
      properties:
        loginBackgroundUrl:
          type: string
          nullable: true
        loginBackgroundVersion:
          type: integer
        loginBackgroundUpdatedAt:
          type: string
          format: date-time
          nullable: true
        appLogoUrl:
          type: string
          nullable: true
        appLogoVersion:
          type: integer
        appLogoUpdatedAt:
          type: string
          format: date-time
          nullable: true
        loginTitle:
          type: string
          nullable: true
        loginMessage:
          type: string
          nullable: true
        assets:
          type: array
          items:
            $ref: '#/components/schemas/BrandingAssetResponse'
    UpdateLoginTextRequest:
      type: object
      properties:
        title:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
    BrandingAssetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
        contentType:
          type: string
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
    HomeVideoCategoryConfigMode:
      type: string
      enum: [ALL, SELECTED]
    HomeVideoCategoryConfigResponse:
      type: object
      properties:
        mode:
          $ref: '#/components/schemas/HomeVideoCategoryConfigMode'
        enabledCategoryIds:
          type: array
          items:
            type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
    HomeVideoCategoryConfigRequest:
      type: object
      required: [mode]
      properties:
        mode:
          $ref: '#/components/schemas/HomeVideoCategoryConfigMode'
        enabledCategoryIds:
          type: array
          items:
            type: string
    FeaturedContentSource:
      type: string
      description: >-
        Logical content source owning the referenced contentId. Only PODCAST
        exists today; new sources are added here as they're onboarded.
      enum: [PODCAST]
    HomePlayerType:
      type: string
      enum: [MINI]
    HomePlayerPosition:
      type: string
      enum: [TOP, BOTTOM]
    HomeFeaturedPlayerConfigResponse:
      type: object
      properties:
        enabled:
          type: boolean
        contentSource:
          allOf:
            - $ref: '#/components/schemas/FeaturedContentSource'
          nullable: true
        contentId:
          type: string
          nullable: true
        playerType:
          $ref: '#/components/schemas/HomePlayerType'
        position:
          $ref: '#/components/schemas/HomePlayerPosition'
        updatedAt:
          type: string
          format: date-time
          nullable: true
    UpdateHomeFeaturedPlayerConfigRequest:
      type: object
      required: [enabled, playerType, position]
      properties:
        enabled:
          type: boolean
        contentSource:
          allOf:
            - $ref: '#/components/schemas/FeaturedContentSource'
          nullable: true
        contentId:
          type: string
          nullable: true
        playerType:
          $ref: '#/components/schemas/HomePlayerType'
        position:
          $ref: '#/components/schemas/HomePlayerPosition'
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time